feathr_project/feathr/api/requirements.txt
azure-core
azure-purview-catalog==1.0.0b2
fastapi
opencensus-ext-azure
pyapacheatlas
pydantic
uvicorn
